with-django/requirements.txt
asgiref==3.9.1
Django==5.2.5
sqlparse==0.5.3
python-decouple==3.8
djangorestframework==3.15.2
